partially implemented a working heap (it was working, but now calling `load_idtr` causes crash)

+++ Sources/Kernel/memory/HeapMemoryBlock.swift
@@ -1,7 +1,7 @@
 
 @safe
 struct HeapMemoryBlock {
-    var size:Int
+    var size:UInt64
     var isFree:Bool
     var next:UnsafeMutablePointer<HeapMemoryBlock>?
 }
